Mv bytecode merklization functions to own file
